+++ RayTracerFramework/RayTracerSmall/MoveCom.h
@@ -5,7 +5,7 @@ class MoveCom
 	int startFrame, endFrame, sphere, amountOfFrames;
 	Vec3f pos, movementPerFrame;
 public:
-	MoveCom(int startFrame, int endFrame, int sphere,int x,int y,int z);
+	MoveCom(int startFrame, int endFrame, int sphere,float x,float y,float z);
 	~MoveCom();
 	int GetSphere();
 	Vec3f GetPosAtFrame(int frameNum);
